Update "hackers" to the more correct term.
See the conversation at https://www.loomio.org/d/SEshIfzW/use-of-the-term-hacker- it was agreed that the term "Hacker" now carries a positive connotation, for example "Hack-a-thons" such as GovHack. The more correct term was to use adversary or cracker and specifically the behaviour they were performing in relation to para. 28 in NZGOAL-SE was malicious exploitation of bugs.

We've made reference to the Internet Security Glossary for a full definition of "Cracker" as this term is often confused with the former term, Hacker.