Update CONTRIBUTING to include clear guidance on pull requests
This is included in the [human readable versions of the revision](https://opendatanz.github.io/nzgoal-se/). It is mostly for the final week following the revisions (last public input before further due process). This is in order for us to review, analyse and consider all the feedback during public consultation.