Improve readability of alternative licence examples
I has trouble understanding what "...to risk but the GPL itself would not allow it in some cases and might not in others. " meant. Removed from above unless we can clarify the mean of it (will revise before merging).

Also, I've swapped out the reference to what the AGPL does. I found the previous misleading as it appeared I needed to actually put the code for download on the server where the software is. Reading over the AGPL, it doesn't require this as far as I can tell.