Minor updates to the security disclosure process
The major update here is to include information about litigation/prosecution of reporters. I think it deserves to be called out especially, as making this less of an issue for white-hat hackers means it's much more likely that they will actually work with you to close the hole, instead of ignoring it.

Perhaps some example wording would help here - for example if you can access e.g. an invoicing system which has unencrypted usernames/passwords stored in it should be reported; and the researcher should not then use those details to try and access other systems. A number of bug bounty programs (e.g. Microsoft, GitHub) have rules similar to this which could help here.

The other changes are minor, and mainly factual (e.g. GICO mis-spelling of GCIO).